+++ generator/page_generators/documentation/panther/value_stages.js
@@ -21,32 +21,32 @@ page.text("All expressions have a value stage which dictate which stage of compi
 
 
 
-page.anchor("comptime");
-page.h2("Comptime");
-search.addSearchTarget("Comptime Value Stage", "documentation/panther/value_stages.html#comptime", [search.Category.PANTHER, search.Category.DOCUMENTATION]);
-page.text("Comptime values are values that are known at compile-time and can be used as comptime expressions.");
-page.text("The following are runtime expressions:");
+page.anchor("constexpr");
+page.h2("Constexpr");
+search.addSearchTarget("Constexpr Value Stage", "documentation/panther/value_stages.html#constexpr", [search.Category.PANTHER, search.Category.DOCUMENTATION]);
+page.text("Constexpr values are values that are known at compile-time and can be used as constexpr expressions.");
+page.text("The following are constexpr expressions:");
 page.bullets([
 	terms.get("def variables"),
 	"template parameters",
-	"comptime function calls with all parameters being comptime"
+	"constexpr function calls with all parameters being constexpr"
 ]);
 
 
 
-page.anchor("constexpr");
-page.h2("Constexpr");
-search.addSearchTarget("Constexpr Value Stage", "documentation/panther/value_stages.html#constexpr", [search.Category.PANTHER, search.Category.DOCUMENTATION]);
-page.text("The term \"constexpr\" in Panther does not have the same exact meaning as it does in C++. Constexpr values are values that may be used inside a comptime function but are not comptime themselves. Anywhere that requires a constexpr value also accept comptime values.");
-page.text("The following are runtime expressions:");
+page.anchor("comptime");
+page.h2("Comptime");
+search.addSearchTarget("Comptime Value Stage", "documentation/panther/value_stages.html#comptime", [search.Category.PANTHER, search.Category.DOCUMENTATION]);
+page.text("Comptime values are values that may be used inside a constexpr function but are not constexpr themselves. Anywhere that requires a comptime value also accept constexpr values.");
+page.text("The following are comptime expressions:");
 page.bullets([
 	"global " + terms.get("const variables"),
-	"parameters in a comptime function",
-	"return parameters in a comptime function",
-	"error parameters in a comptime function",
-	terms.get("const variables") + " defined in a comptime function",
-	terms.get("var variables") + " defined in a comptime function",
-	"comptime function calls with all parameters being comptime and/or constexpr (must have at least one constexpr parameter, or else it's comptime)",
+	"parameters in a constexpr function",
+	"return parameters in a constexpr function",
+	"error parameters in a constexpr function",
+	terms.get("const variables") + " defined in a constexpr function",
+	terms.get("var variables") + " defined in a constexpr function",
+	"constexpr function calls with all parameters being constexpr and/or comptime (must have at least one comptime parameter, or else it's constexpr)",
 ]);
 
 
@@ -54,7 +54,7 @@ page.bullets([
 page.anchor("runtime");
 page.h2("Runtime");
 search.addSearchTarget("Runtime Value Stage", "documentation/panther/value_stages.html#runtime", [search.Category.PANTHER, search.Category.DOCUMENTATION]);
-page.text("Runtime values are values that may only be used at runtime as they are not known and/or cannot be calculated at compile-time. Anywhere that that requires a runtime value also accept constexpr and comptime values.");
+page.text("Runtime values are values that may only be used at runtime as they are not known and/or cannot be calculated at compile-time. Anywhere that that requires a runtime value also accept comptime and constexpr values.");
 page.text("The following are runtime expressions:");
 page.bullets([
 	"global " + terms.get("var variables"),